creating course outline bug fixed and searching table element functionality seperated by a custom hook to reduce code duplication
